+++ packages/terra-clinical-item-display/src/terra-dev-site/doc/example/TextStyles.jsx
@@ -4,7 +4,8 @@ import ItemDisplay from 'terra-clinical-item-display/lib/ItemDisplay';
 
 const component = () => (
   <React.Fragment>
-    <ItemDisplay text="Primary" />
+    <ItemDisplay text="Default (same as Primary)" />
+    <ItemDisplay text="Primary" textStyle="primary" />
     <ItemDisplay text="Strong" textStyle="strong" />
     <ItemDisplay text="Attention" textStyle="attention" />
     <ItemDisplay text="Secondary" textStyle="secondary" />