+++ packages/terra-clinical-item-display/src/terra-dev-site/doc/example/TextStylesDisabled.jsx
@@ -4,7 +4,8 @@ import ItemDisplay from 'terra-clinical-item-display/lib/ItemDisplay';
 
 const component = () => (
   <React.Fragment>
-    <ItemDisplay text="Primary" isDisabled />
+    <ItemDisplay text="Default" isDisabled />
+    <ItemDisplay text="Primary" textStyle="primary" isDisabled />
     <ItemDisplay text="Strong" textStyle="strong" isDisabled />
     <ItemDisplay text="Attention" textStyle="attention" isDisabled />
     <ItemDisplay text="Secondary" textStyle="secondary" isDisabled />