+++ packages/terra-clinical-item-display/src/terra-dev-site/test/clinical-item-display/display/TextStyleItemDisplay.test.jsx
@@ -4,6 +4,7 @@ import ItemDisplay from '../../../../ItemDisplay';
 
 const views = () => (
   <React.Fragment>
+    <ItemDisplay text="test text (unset textStyle)" id="test-display-undefined" />
     <ItemDisplay text="test text" textStyle="primary" id="test-display-primary" />
     <ItemDisplay text="test text" textStyle="strong" id="test-display-strong" />
     <ItemDisplay text="test text" textStyle="attention" id="test-display-attention" />